refactor(review): extract JSON parsing utils from llm-client (#292)
Move extractJSONFromCodeBlock and estimatePromptTokens into a dedicated
json-utils module to decouple pure string helpers from the HTTP client.

- Delete dead code: parseJSONResponse, filterStringValues (exported but
  never imported)
- Remove duplicate estimatePromptTokens from openrouter.ts
- Rename test file to match new module, add estimatePromptTokens tests
